# JSON: Fix complex64 encoding precision